created a package for the domain logic, gui logic, and tests. Also updated the README
