--- src/Player.java
@@ -1,4 +1,0 @@
-
-public class Player {
-
-}
