--- src/Tax.java
@@ -1,4 +1,0 @@
-
-public class Tax {
-
-}
