+++ src/a4/domain/OpenSpace.java
@@ -1,3 +1,4 @@
+package cs414.c.a4.domain;
 
 public class OpenSpace {
