+++ src/a4/domain/Street.java
@@ -1,0 +1,5 @@
+package cs414.c.a4.domain;
+
+public class Street {
+
+}
