+++ src/a4/gui/Frame.java
@@ -1,0 +1,5 @@
+package cs414.c.a4.gui;
+
+public class Frame {
+
+}
